requirements.txt
Django==2.1.8
django-mdeditor==0.1.13
Markdown==3.1
numpy==1.16.2
Pillow==6.0.0
Pygments==2.3.1
pytz==2018.9
wordcloud==1.5.0
